# src/main/java/com/mikels/citasalud/domain/model/Medico.java
package com.mikels.citasalud.domain.model;

import java.util.UUID;

import lombok.Builder;
import lombok.Getter;

@Getter
@Builder
public class Medico {

    private final UUID id;
    private final String nombre;
    private final String especialidad;
    private final boolean activo;
}